refactor: update icon component tm-21
