--- apps/frontend/src/libs/components/icon/libs/enums/enums.ts
@@ -1,1 +1,0 @@
-export { IconName } from "./icon-name.enum.js";
