--- apps/frontend/src/libs/components/icon/libs/enums/icon-name.enum.ts
@@ -1,5 +1,0 @@
-const IconName = {
-	HOME: "home",
-} as const;
-
-export { IconName };
