+++ apps/frontend/src/libs/components/icon/libs/maps/maps.ts
@@ -1,0 +1,1 @@
+export { iconNameToSvg } from "./icon-name-to-svg.map.js";
