+++ apps/frontend/src/libs/components/icon/libs/types/icon-name.type.ts
@@ -1,0 +1,3 @@
+type IconName = "home";
+
+export { IconName };
