+++ apps/frontend/src/libs/components/icon/libs/types/types.ts
@@ -1,0 +1,1 @@
+export { type IconName } from "./icon-name.type.js";
